Add copy construction.

+++ rebuild.sh
@@ -9,7 +9,8 @@ then
 fi
 I=1
 # can remove once we have compiler hash
-rm -rf .obj
+# see https://stackoverflow.com/questions/3601515/how-to-check-if-a-variable-is-set-in-bash
+if [ -z "${FAST+x}" ]; then rm -rf .obj; fi
 build/cx $FLAGS -Pcompiler:build/src -Pnext:src src/main.cx -o build/cx_test$I
 SUM="$(objdump -S build/cx_test$I |grep -v file\ format |md5sum)"
 SUMNEXT=""
@@ -17,7 +18,7 @@ while true
 do
     K=$((I+1))
     # can remove once we have compiler hash
-    rm -rf .obj
+    if [ -z "${FAST+x}" ]; then rm -rf .obj; fi
     build/cx_test$I $FLAGS -Pcompiler:src src/main.cx -o build/cx_test$K
     SUMNEXT="$(objdump -S build/cx_test$K |grep -v file\ format |md5sum)"
     if [ "$SUM" == "$SUMNEXT" ]; then break; fi
